backend-schedule-main/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Werkzeug==3.1.3

# Database
psycopg2-binary==2.9.9
alembic==1.13.1
SQLAlchemy==2.0.23

# Environment variables
python-dotenv==1.1.1

# Google AI (for Gemini)
google-generativeai==0.3.2

# Utilities
click==8.2.1
python-dateutil==2.9.0.post0
pytz==2025.2

# Development
blinker==1.9.0
colorama==0.4.6